aider: Use `hub` command to create pull requests in the workflow file.
# Aider chat conversation:

USER: There is no PR created in the repo for unmerged changes. I think we need to try another way of creating the PR

ASSISTANT: To address the issue of the PR not being created, we can try using a different action for creating pull requests. One such action is `hub`, which is a command-line tool that wraps `git` in order to extend it with extra features and commands that make working with GitHub easier.

Let's modify the `.github/workflows/openfaas-pre-docker-build-pr.yaml` file to use `hub` for creating the pull request.

Here is the diff for the changes:

```diff
--- .github/workflows/openfaas-pre-docker-build-pr.yaml
+++ .github/workflows/openfaas-pre-docker-build-pr.yaml
@@ ... @@
    - name: Create or update PR with new release number
-     uses: peter-evans/create-pull-request@v4
-     with:
-       token: ${{ secrets.GITHUB_TOKEN }}
-       commit-message: "Prepare for Docker build version ${{ steps.versioning.outputs.new_version }}"
-       title: "Docker Build Version ${{ steps.versioning.outputs.new_version }}"
-       body: |
-         ## Changelog
-         <!-- markdown-changelog-generator -->
-       branch: "release-${{ steps.versioning.outputs.new_version }}"
-       base: "main"
-       labels: "pre-release"
-       # update-branch: true
-       # title-regex: "Docker Build Version"
-       # body-includes-regex: "markdown-changelog-generator"
+     run: |
+       git config --global user.email "action@github.com"
+       git config --global user.name "GitHub Action"
+       git checkout -b "release-${{ steps.versioning.outputs.new_version }}"
+       git commit --allow-empty -m "Prepare for Docker build version ${{ steps.versioning.outputs.new_version }}"
+       git push origin "release-${{ steps.versioning.outputs.new_version }}"
+       echo "## Changelog" > PR_BODY.md
+       echo "<!-- markdown-changelog-generator -->" >> PR_BODY.md
+       hub pull-request -f -p -b main -F PR_BODY.md -l "pre-release"
+     env:
+       GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
```

This change replaces the `peter-evans/create-pull-request` action with a series of `git` and `hub` commands to create the pull request. The `hub` command is used with the `-f` flag to force the creation of the pull request, `-p` to push the branch, `-b` to specify the base branch, `-F` to read the pull request message from a file, and `-l` to add a label to the pull request. The `GITHUB_TOKEN` is provided as an environment variable to authenticate the `hub` command with GitHub.